Get enviPath objects

epGet returns raw objects from enviPath.

epGet(type, init = NULL, pkg = NULL, property = NULL)

Arguments

type

Character scalar. String specifying the object type to be fetched.

init

Character vector. Vector of strings specifying the initial values of type that should be fetched. When null, all objects belonging to type in pkg are used. (Default: NULL)

pkg

Character scalar. String specifying the unique identifier of the package from which objects should be mapped When null, EAWAG-BBD is used. (Default: NULL)

property

Character scalar. String specifying the property to be fetched from the objects. When null, all elements of the objects are returned. (Default: NULL)

Value

A list of objects

Examples

#> Hi nature lover, welcome to enviPath!
rxn_id <- "2b6bbcc5-77f4-4bed-92a9-731cdc978f6a"

epGet("reaction", rxn_id)
#> $aliases
#> list()
#> 
#> $description
#> [1] "no description"
#> 
#> $ecNumbers
#> list()
#> 
#> $educts
#>             compoundName
#> 1 4-Hydroxybutyraldehyde
#>                                                                                                                                                               id
#> 1 https://envipath.org/package/32de3cf4-e3e6-4168-956e-32fa5ddb0ce1/compound/d2c48ebe-5a5a-4f5b-b42e-8ff23c873f6c/structure/7281d994-9f4a-4f7f-9e50-90b131c11da3
#>      smiles
#> 1 C(CCO)C=O
#> 
#> $id
#> [1] "https://envipath.org/package/32de3cf4-e3e6-4168-956e-32fa5ddb0ce1/reaction/2b6bbcc5-77f4-4bed-92a9-731cdc978f6a"
#> 
#> $identifier
#> [1] "reaction"
#> 
#> $medlineRefs
#> list()
#> 
#> $multistep
#> [1] FALSE
#> 
#> $name
#> [1] "Eawag BBD reaction r0014"
#> 
#> $pathways
#>                                                                                                               id
#> 1 https://envipath.org/package/32de3cf4-e3e6-4168-956e-32fa5ddb0ce1/pathway/c3d1b8d5-9ffb-49b8-a1bf-bf3d7dfd1ac3
#>              name reviewStatus identifier
#> 1 Tetrahydrofuran     reviewed    pathway
#> 
#> $products
#>         compoundName
#> 1 4-Hydroxybutanoate
#>                                                                                                                                                               id
#> 1 https://envipath.org/package/32de3cf4-e3e6-4168-956e-32fa5ddb0ce1/compound/a41e6b76-b462-49d4-ae42-391ef00bf98c/structure/befe8eeb-e8f8-4b4a-90b1-d0584a75f8cb
#>            smiles
#> 1 C(CC(=O)[O-])CO
#> 
#> $references
#> $references$rheaReferences
#> list()
#> 
#> $references$uniprotCount
#> list()
#> 
#> 
#> $reviewStatus
#> [1] "reviewed"
#> 
#> $scenarios
#> list()
#> 
#> $smirks
#> [1] "C(CCO)C=O>>C(CC(=O)[O-])CO"
#>