List enviPath objects

epList returns a list with info on objects available in enviPath.

epList(type, pkg = NULL)

Arguments

type

Character scalar. String specifying the object type to listed.

pkg

Character scalar. String specifying the unique identifier of the package from which objects should be listed. When null, all available packages are used. (Default: NULL)

Value

A data frame with all objects belonging to type in pkg.

Examples

# Perform login
epLogin(username, password)
#> Hi nature lover, welcome to enviPath!

# List packages
pkg_df <- epList("package")

# Select reviewed packages
to_keep <- pkg_df$reviewStatus == "reviewed"
pkg_df <- pkg_df[to_keep, ]

# Remove lengthy description
pkg_df$description <- NULL

# View some packages
head(pkg_df)
#>                        name                                   id reviewStatus
#> 3  Public Prediction Models 134886bb-b52e-4ad6-91d3-a02cecde6d95     reviewed
#> 6            EAWAG-SEDIMENT f05e38d8-e9b4-4c3e-b0d8-9ab29966eccf     reviewed
#> 7             enviPath-PFAS 87a49584-d937-482c-9c33-25928dcb02a8     reviewed
#> 9                EAWAG-SOIL 5882df9c-dae1-4d80-a40e-db4724271456     reviewed
#> 11             EAWAG-SLUDGE 521c547a-fd2a-491c-ad5b-7eaa1577fb65     reviewed
#> 14                EAWAG-BBD 32de3cf4-e3e6-4168-956e-32fa5ddb0ce1     reviewed
#>    Pathways Rules Compounds Reactions Relative Reasoning Scenarios
#> 3         0     0         0         0                  3         0
#> 6       179     0       810       961                  0      1622
#> 7       214     0       590       915                  0       288
#> 9       317     0      2608      2445                  0      4881
#> 11      183     0      1067       494                  0        85
#> 14      219   499      1399      1480                  0      1915

# Select desired package
pkg_name <- "EAWAG-BBD"
pkg_id <- pkg_df$id[pkg_df$name == pkg_name]

# List pathways from desired package
path_df <- epList("pathway", pkg = pkg_id)

# View some pathways
head(path_df)
#>                      name                                   id reviewStatus
#> 1                 Uranium 259b990e-d653-4201-9282-04f2c6b05ca0     reviewed
#> 2 2-Aminobenzenesulfonate 1e7066c2-ff45-44f4-9086-e37db18400df     reviewed
#> 3                Nicotine 1e94c7d7-7f3a-42c7-9dd0-eb1ab217d7ec     reviewed
#> 4         2-Aminobenzoate dfc8c935-9d73-4067-bfe6-db3501d2075f     reviewed
#> 5                Caffeine 1d537657-298c-496b-9e6f-2bec0cbe0678     reviewed
#> 6    Dimethylisophthalate a0641e14-3811-482f-a366-cc53f893fce4     reviewed