List enviPath objects

epList returns a list with info on objects available in enviPath.

epList(type, pkg = NULL)

Arguments

type

Character scalar.

pkg

Character scalar.

Value

A data frame.

Examples

#> Hi RiboRings, welcome to enviPath!
epList("package")
#>                                                                     name
#> 1  Autogenerated Package for Batch Prediction 2026-09-01 00:04:40.517620
#> 2                                                     RiboRings’ Package
#> 3                                               Public Prediction Models
#> 4  Autogenerated Package for Batch Prediction 2026-09-01 00:06:57.531194
#> 5                                   32de3cf4-e3e6-4168-956e-32fa5ddb0ce1
#> 6                                                         EAWAG-SEDIMENT
#> 7                                                          enviPath-PFAS
#> 8  Autogenerated Package for Batch Prediction 2026-09-01 00:04:22.775931
#> 9                                                             EAWAG-SOIL
#> 10 Autogenerated Package for Batch Prediction 2026-09-01 00:08:26.613909
#> 11                                                          EAWAG-SLUDGE
#> 12                                  32de3cf4-e3e6-4168-956e-32fa5ddb0ce1
#> 13                                               enviPath-PFAS_temporary
#> 14                                                             EAWAG-BBD
#> 15                                                    enviPath-PFAS-2025
#>                                      id reviewStatus
#> 1  5b594bf6-56a8-4ce5-9a63-5fd4b525756e   unreviewed
#> 2  aee2524b-79f3-4640-a519-aac543ac05e3   unreviewed
#> 3  134886bb-b52e-4ad6-91d3-a02cecde6d95     reviewed
#> 4  7a3bddd7-c1a2-4699-a885-8b9044fd4fe2   unreviewed
#> 5  d1898032-350b-438a-a702-63906c7ed76a   unreviewed
#> 6  f05e38d8-e9b4-4c3e-b0d8-9ab29966eccf     reviewed
#> 7  87a49584-d937-482c-9c33-25928dcb02a8     reviewed
#> 8  995230c9-167a-4108-8002-d8fadd242dff   unreviewed
#> 9  5882df9c-dae1-4d80-a40e-db4724271456     reviewed
#> 10 c3c6699a-8c78-4987-af4d-4a041ffc448d   unreviewed
#> 11 521c547a-fd2a-491c-ad5b-7eaa1577fb65     reviewed
#> 12 7245f711-51d0-4b1e-b49d-99a9d9744c96   unreviewed
#> 13 eb601621-3b32-4b64-b84a-97fdaeb366fb   unreviewed
#> 14 32de3cf4-e3e6-4168-956e-32fa5ddb0ce1     reviewed
#> 15 d2cfb5af-4ea0-4375-9a48-f2e776e44636   unreviewed
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                description
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  This Package was generated automatically for the batch prediction task.
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          This package was generated during registration.
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Package to make Prediction Models publicly available
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  This Package was generated automatically for the batch prediction task.
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   string
#> 6                                                                                                                                                                                                                                                                                          The "EAWAG-SEDIMENT" package contains pathways and kinetic data from biodegradation studies performed in water-sediment systems. This information is extracted from pesticide registration dossiers (draft assessment reports, DAR) made publically available through the European Food Safety Authority (<a href="https://www.efsa.europa.eu/en/calls/consultations" target="_blank">EFSA</a>). The package also contains information on different experimental conditions that are stored in the object <a href="https://envipath.org/package/f05e38d8-e9b4-4c3e-b0d8-9ab29966eccf/scenario" target="_blank">scenario</a>. <a href="https://envipath.org/package/f05e38d8-e9b4-4c3e-b0d8-9ab29966eccf/compound" target="_blank">Compounds</a> in the pathway are associated with a given <a href="https://envipath.org/package/f05e38d8-e9b4-4c3e-b0d8-9ab29966eccf/scenario" target="_blank">scenario</a> if they have been observed under the specific experimental conditions. If available, a total system biotransformation half-life (DT50) value is associated with a given compound in the pathway and a specific scenario, along with dissipation half-lives from the water and the sediment phases.
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     A collection of literature-reported biotransformation pathways for per- and polyfluorinated alkyl substances (PFASs). Last updated on June 9th 2026.
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  This Package was generated automatically for the batch prediction task.
#> 9  "EAWAG-SOIL" package contains pathway information from soil degradation studies, stored in a biotransformation reaction scheme in the object <a href="https://envipath.org/package/5882df9c-dae1-4d80-a40e-db4724271456/pathway" target="_blank">pathway</a>. This information is extracted from pesticide registration dossiers (draft assessment reports, DAR) made publically available through the European Food Safety Authority (<a href="https://www.efsa.europa.eu/en/calls/consultations" target="_blank">EFSA</a>). The package also contains information on different experimental conditions that are stored in the object <a href="https://envipath.org/package/5882df9c-dae1-4d80-a40e-db4724271456/scenario" target="_blank">scenario</a>. <a href="https://envipath.org/package/5882df9c-dae1-4d80-a40e-db4724271456/compound" target="_blank">Compounds</a> in the pathway are associated with a given <a href="https://envipath.org/package/5882df9c-dae1-4d80-a40e-db4724271456/scenario" target="_blank">scenario</a> if they have been observed under the specific experimental conditions. If available, a biotransformation half-life (DT50) value is additionally associated with a given compound in the pathway and a specific scenario. More details concerning enviPath conventions that were used to store specific data from DAR, and their description can be found in our <a href="https://wiki.envipath.org/doku.php?id=packages#eawag-soil" target="_blank">wiki</a>.
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 This Package was generated automatically for the batch prediction task.
#> 11                                                                                                                                                                                                         The <a href="https://envipath.org/package/521c547a-fd2a-491c-ad5b-7eaa1577fb65">EAWAG-SLUDGE</a> package contains pathways and kinetic information about microbially mediated transformation processes in biological wastewater treatment. The information was extracted from scientific publications in different journals. The biotransformation reaction schemes are stored as <a href="https://envipath.org/package/4a3cd0f4-4d2b-4f00-b3e6-a29e721f7038/pathway">pathways</a>. The experimental conditions and additional information (e. g. sludge source) are collected in the <a href="https://envipath.org/package/4a3cd0f4-4d2b-4f00-b3e6-a29e721f7038/scenario">scenario</a> objects. Any observed transformation products (<a href="https://envipath.org/package/4a3cd0f4-4d2b-4f00-b3e6-a29e721f7038/compound">compounds</a>) are linked to their corresponding experimental metadata (<a href="https://envipath.org/package/4a3cd0f4-4d2b-4f00-b3e6-a29e721f7038/scenario">scenario</a>). Kinetic parameters (rate constants and/or half-lives) including corresponding scenarios are indicated for compounds for which such data is available. More details regarding enviPath conventions can be found in our <a href="https://wiki.envipath.com/index.php/Main_Page">wiki</a>.
#> 12                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  string
#> 13                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             This is a package designed for non-peer-reviewed pathways. Once the pathways are peer-reviewed, this package will be merged into the enviPath-PFAS package.
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Migration package for pathways, reactions and compounds from the EAWAG-BBD database
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  A collection of literature-reported biotransformation pathways for per- and polyfluorinated alkyl substances (PFASs). This version has been replaced by a new version.
#>    Pathways Rules Compounds Reactions Relative Reasoning Scenarios
#> 1         1     0        24        38                  0         0
#> 2         4     0        33        66                  0         0
#> 3         0     0         0         0                  3         0
#> 4         0     0         0         0                  0         0
#> 5         0     0         0         0                  0         0
#> 6       179     0       810       961                  0      1622
#> 7       214     0       590       915                  0       288
#> 8         0     0         0         0                  0         0
#> 9       317     0      2608      2445                  0      4881
#> 10        2     0        21        23                  0         0
#> 11      183     0      1067       494                  0        85
#> 12        0     0         0         0                  0         0
#> 13        3     0        13        11                  0         3
#> 14      219   499      1399      1480                  0      1915
#> 15      124     0       468       804                  0       156

epList("pathway", pkg = "32de3cf4-e3e6-4168-956e-32fa5ddb0ce1")
#>                                                                  name
#> 1                                                             Uranium
#> 2                                             2-Aminobenzenesulfonate
#> 3                                                            Nicotine
#> 4                                                     2-Aminobenzoate
#> 5                                                            Caffeine
#> 6                                                Dimethylisophthalate
#> 7                                                    4-Chlorobiphenyl
#> 8                                                 Toluene (anaerobic)
#> 9                                                           Ibuprofen
#> 10                                                2,4-Dichloroaniline
#> 11                                                       Adamantanone
#> 12                                                   Mordant Yellow 3
#> 13                                                        Caprolactam
#> 14                           Dimethyl Sulfoxide & Organosulfide Cycle
#> 15                                        Organosilicone (an/aerobic)
#> 16                                       Phenanthrene (fungal 9R,10R)
#> 17                                                         Permethrin
#> 18                                                Anthracene (fungal)
#> 19                                            2-Mercaptobenzothiazole
#> 20                                                          Isooctane
#> 21                                         N-Methylmorpholine-N-oxide
#> 22                                                  Trichloroethylene
#> 23                                             3-Nitrotyrosine Family
#> 24                                                           Nitrofen
#> 25                                                 Picric Acid Family
#> 26                                                           Limonene
#> 27                                                      Organomercury
#> 28                                                           Atrazine
#> 29                                    1,5-Anhydro-D-fructose (Fungal)
#> 30                                                            Toluene
#> 31                                                 Tetrachlorobenzene
#> 32                                            Cyclopropanecarboxylate
#> 33                                                          Propylene
#> 34                                               Z-Phenylacetaldoxime
#> 35                                                        Cyclohexane
#> 36                                                           o-Xylene
#> 37                                                            Phorate
#> 38                                                          Carbazole
#> 39                                             N-Nitrosodimethylamine
#> 40                                                 2,4-Dinitrotoluene
#> 41                                           Methionine and Threonine
#> 42                                                         Fosfomycin
#> 43                                                Nitrate (anaerobic)
#> 44                                                       Testosterone
#> 45                                                      Thiobenzamide
#> 46                                                 Phenol (anaerobic)
#> 47                                               n-Hexane (anaerobic)
#> 48              1,1,1-Trichloro-2,2-bis-(4`-chlorophenyl)ethane (DDT)
#> 49                                                         Pyridoxine
#> 50                                                    Dichloromethane
#> 51                                                         Anthracene
#> 52                                            Perchlorate (anaerobic)
#> 53                                                     Imidosuccinate
#> 54                                                            Styrene
#> 55                                                Toluene-4-sulfonate
#> 56                                                           Fluorene
#> 57                                              Valanimycin Synthesis
#> 58                                                           Selenate
#> 59                                                           Tyrosine
#> 60                                                     2-Nitropropane
#> 61                                                          Iprodione
#> 62                                             N-Cyclohexylisocyanide
#> 63                                    2-Chloro-N-isopropylacetanilide
#> 64                                                           Atenolol
#> 65                                            3,3`-Dithiodipropionate
#> 66                                      Trifluoroacetate (An/Aerobic)
#> 67                                                        (+)-Camphor
#> 68                                           Ethylbenzene (anaerobic)
#> 69                                                    Dodecyl sulfate
#> 70                                                          Carbaryl 
#> 71                                                   Osmium Reduction
#> 72                                                          Mandelate
#> 73                                                           Atropine
#> 74                                                           Geraniol
#> 75                                                         Glyphosate
#> 76                                       Pentaerythritol tetranitrate
#> 77                                                2-Methylnaphthalene
#> 78                                                           Furfural
#> 79                                                       Benzonitrile
#> 80                                                           p-Cymene
#> 81                                                         Metamitron
#> 82                                    2-Methylnaphthalene (anaerobic)
#> 83                                                           Tetralin
#> 84                                Hexadecyltrimethylammonium Chloride
#> 85                                                Cyclohexylsulfamate
#> 86                                                      Nitroglycerin
#> 87                                                        Isoproturon
#> 88                                                         Carbofuran
#> 89                            gamma-1,2,3,4,5,6-Hexachlorocyclohexane
#> 90                                  2,4,6-Trinitrotoluene (anaerobic)
#> 91                                  s-Triazine Metabolism Metapathway
#> 92                beta-1,2,3,4,5,6-Hexachlorocyclohexane (an/aerobic)
#> 93                                                        1,4-Dioxane
#> 94                                                        Quinaldine 
#> 95                                                      Thioacetamide
#> 96                                              1,2,3-Tribromopropane
#> 97                                                       Nitrobenzene
#> 98                                              4-Hydroxyacetophenone
#> 99                                                         Morpholine
#> 100                                                         Isoniazid
#> 101                                  Dibenzothiophene Desulfurization
#> 102                                                      alpha-Pinene
#> 103                                              Benzoate (anaerobic)
#> 104                                                         Quercetin
#> 105                                  Carbon Tetrachloride (anaerobic)
#> 106                                                     Chromium (VI)
#> 107                                                        Triazophos
#> 108                                              3-Chloroacrylic Acid
#> 109                                                    Methanogenesis
#> 110                                                     Arsonoacetate
#> 111                                                       Dichlobenil
#> 112                                                           Orcinol
#> 113                                                        Ketoprofen
#> 114                                                    Benzyl Sulfide
#> 115                                                    Acenaphthylene
#> 116                                                      Ethylbenzene
#> 117                                                          Alachlor
#> 118                                                  1-Naphthoic Acid
#> 119                                                     Phenylacetate
#> 120                                                    Methylarsonate
#> 121                                           Methyl tert-butyl ether
#> 122                                     4-Fluorobenzoate (an/aerobic)
#> 123                                       Naphthalenesulfonate Family
#> 124                                                     Cyanuric Acid
#> 125                                                         Malathion
#> 126                                                     Chlorobenzene
#> 127                                                1,2-Dichloroethane
#> 128                                                          Biphenyl
#> 129                                               Dimethylphosphinate
#> 130                                             Tetrabromobisphenol A
#> 131                                                          Diazepam
#> 132                                                       Bisphenol A
#> 133                                                      Triphenyltin
#> 134                                                       Thiocyanate
#> 135                                                       Naphthalene
#> 136                           Branched Chain Dodecylbenzene Sulfonate
#> 137                                                3-Phenylpropionate
#> 138                               1,2,4-Trichlorobenzene (an/aerobic)
#> 139                                   Nitrophenol Family (an/aerobic)
#> 140                                                     Methamidophos
#> 141                           Hexahydro-1,3,5-trinitro-1,3,5-triazine
#> 142                             4-(1-Ethyl-1,4-dimethyl-pentyl)phenol
#> 143                                 1-Aminocyclopropane-1-carboxylate
#> 144                                                5-Nitroanthanilate
#> 145                                                    Phenol Family 
#> 146                                               Methyl ethyl ketone
#> 147                                                   Tetrahydrofuran
#> 148                                             Pyrrole-2-carboxylate
#> 149                                         Technetium Immobilization
#> 150                                               1-Methylnaphthalene
#> 151                                          Pivalic Acid (anaerobic)
#> 152                                                          m-Cresol
#> 153                                                          n-Octane
#> 154                                                   Methyl fluoride
#> 155 1,1,1-Trichloro-2,2-bis-(4`-chlorophenyl)ethane (DDT) (anaerobic)
#> 156                                                  Dibenzo-p-dioxin
#> 157                                                          p-Xylene
#> 158                                         3-(4-Sulfophenyl)butyrate
#> 159                                       2-Aminobenzoate (anaerobic)
#> 160                                                C1 Metabolic Cycle
#> 161                                                          Vanillin
#> 162                                                      Isocarbophos
#> 163                                    Benzothiophene Desulfurization
#> 164                                                         Syringate
#> 165                                              Methanesulfonic Acid
#> 166                                           1,3-Dichloro-2-propanol
#> 167                                               1,4-Dichlorobenzene
#> 168                                                           Thiamin
#> 169                                           Bromoxynil (an/aerobic)
#> 170                                                         Cinnamate
#> 171                                           Thiophene-2-carboxylate
#> 172                                    Nitrilotriacetate (an/aerobic)
#> 173                                                      Cypermethrin
#> 174                                                      Phenanthrene
#> 175                                                   Trinitrotoluene
#> 176                                                     Acrylonitrile
#> 177                                                    Nylon Oligomer
#> 178                                      Phenanthrene (fungal 9S,10S)
#> 179               Hexahydro-1,3,5-Trinitro-1,3,5-Triazine (anaerobic)
#> 180                                                 3-Methylquinoline
#> 181                                                           Benomyl
#> 182                                                   Starch Oligomer
#> 183                                                       Bisphenol F
#> 184                                                           Cocaine
#> 185                                 2,4,5-Trichlorophenoxyacetic Acid
#> 186                                      Dibenzothiophene Degradation
#> 187                                                        Endosulfan
#> 188                                          N-Oleoyl-N-methyltaurine
#> 189                                                         Acetylene
#> 190                                                    Dimethyl ether
#> 191                                1,1,1-Trichloroethane (an/aerobic)
#> 192                                                   o-Nitrobenzoate
#> 193                                               Menthol (anaerobic)
#> 194                                                    Tri-n-butyltin
#> 195                                                 4-Hydroxypyridine
#> 196                                                          m-Xylene
#> 197                                          Pentachlorophenol Family
#> 198                                            Parathion (an/aerobic)
#> 199                                               1,3-Dichloropropene
#> 200                                                            Asulam
#> 201                                                         Cyanamide
#> 202                                    2,4-Dichlorophenoxyacetic Acid
#> 203                                                       Citronellol
#> 204                                                    Benzo[a]pyrene
#> 205                                                 Caffeine (fungal)
#> 206                                               Gallate (anaerobic)
#> 207                                                  Phthalate Family
#> 208                                                   Triethanolamine
#> 209                                                      Dibenzofuran
#> 210                                                          Benzoate
#> 211                                                            Pyrene
#> 212                                                     Hypophosphite
#> 213                                              2,4-Dichlorobenzoate
#> 214                                               2,4-Dichlorotoluene
#> 215                                     Tetrachloroethene (anaerobic)
#> 216                                                  3-Fluorobenzoate
#> 217                                             N-Cyclopropylmelamine
#> 218                                      4-Carboxy-4`-Sulfoazobenzene
#> 219                                           N,N-Diethyl-m-toluamide
#>                                       id reviewStatus
#> 1   259b990e-d653-4201-9282-04f2c6b05ca0     reviewed
#> 2   1e7066c2-ff45-44f4-9086-e37db18400df     reviewed
#> 3   1e94c7d7-7f3a-42c7-9dd0-eb1ab217d7ec     reviewed
#> 4   dfc8c935-9d73-4067-bfe6-db3501d2075f     reviewed
#> 5   1d537657-298c-496b-9e6f-2bec0cbe0678     reviewed
#> 6   a0641e14-3811-482f-a366-cc53f893fce4     reviewed
#> 7   3d627cfe-a944-431e-afee-1d92014bb035     reviewed
#> 8   433fdde8-2f07-49d3-acc0-47aa695d2c9a     reviewed
#> 9   ec61fd05-6400-4ab1-8d9c-56fa16cc035e     reviewed
#> 10  b83729c6-d40a-477d-8618-a05000ca4f1a     reviewed
#> 11  b91e3577-f69f-464a-9b0b-5d9767864b6e     reviewed
#> 12  43e85f4c-7efe-4a45-8f51-e7f25e4ed16a     reviewed
#> 13  ac2a61a5-22e2-47a4-8a44-b3a027c90428     reviewed
#> 14  439c86fc-abb3-4269-b158-e166fbc3a974     reviewed
#> 15  4fb5b3e7-dd23-44dd-9d9d-ae9c6ce4a921     reviewed
#> 16  a2a860c7-8ae2-44be-b979-84e11c1d5bd6     reviewed
#> 17  90e662ba-96a7-4e37-8872-76a146717dc3     reviewed
#> 18  cb600409-f044-4e55-890c-495269817d55     reviewed
#> 19  7f206519-4fab-488f-bb6f-48b8cbbb96af     reviewed
#> 20  d2157424-aef9-4997-bbb0-f8678b23d3ec     reviewed
#> 21  16452903-5359-4790-8909-998e8ce19708     reviewed
#> 22  956965db-6108-43b1-a8c3-e7c7d57ae291     reviewed
#> 23  8296403d-597b-4c86-b888-b9827dac5fa5     reviewed
#> 24  53c7dbcf-133d-4fad-a5bf-f4ed832a9ab6     reviewed
#> 25  6598285e-566b-4f53-8b72-a2c84d9ed0f6     reviewed
#> 26  79f20508-191a-4631-b636-05cdf806e51e     reviewed
#> 27  07f4cb0e-c553-4eda-a826-9f754b12eb62     reviewed
#> 28  b3d4b253-37e4-4aac-929f-83bd59dc5277     reviewed
#> 29  e20b5028-7865-4ae0-b512-25dca0a10603     reviewed
#> 30  b0cb0262-3f5d-4a17-81ff-1560f9f5b4cd     reviewed
#> 31  d0ca2469-a571-45a4-9a2e-7bc5a09e87b8     reviewed
#> 32  cb7f1a50-be0e-4b71-8b08-70c8ee29e1d7     reviewed
#> 33  0b970043-0d9d-44b5-af68-1430580caef1     reviewed
#> 34  a580c7dd-d9b2-4ec7-b70b-84b828ccdf1d     reviewed
#> 35  8c3caf9b-5b5d-4173-b2c4-0db8c3d89a07     reviewed
#> 36  541a1061-2fe7-418c-84ca-4a07f3c8cf0c     reviewed
#> 37  b11f4d8c-d957-4822-91f0-d521026fc174     reviewed
#> 38  853d5810-7f5c-4e43-8fe6-9cc241a2b30f     reviewed
#> 39  a78b4206-c81e-4169-bd22-9d3f8a266030     reviewed
#> 40  b9164e5d-03aa-4d62-b5fb-537435f832c2     reviewed
#> 41  5e7c8a50-0bf7-4cc3-9e88-7bffce1e886c     reviewed
#> 42  9a53d7d0-c41e-448f-bdf3-ad6cec6786f3     reviewed
#> 43  b31e0bd5-2357-4497-b221-f307fd6245e8     reviewed
#> 44  47d8d0a3-cc85-4460-afcb-fbe79e595e53     reviewed
#> 45  85e7f0d7-0515-434d-ad3d-1aa9682fef30     reviewed
#> 46  c1dfcc66-f974-4cdc-a026-cc1274518f3f     reviewed
#> 47  1cee659a-fdf9-4433-b285-00dbfcc27bca     reviewed
#> 48  3f58e4d4-1c63-4b30-bf31-7ae4b98899fe     reviewed
#> 49  84d54a87-173a-49f9-a87a-f43d35ef6a42     reviewed
#> 50  a2e77e6c-4591-4680-aac4-4ad04321d736     reviewed
#> 51  fe417a7c-671a-476b-9fe0-0e73133aa4db     reviewed
#> 52  63f713de-fbe6-441e-a1a1-27c26c9920b8     reviewed
#> 53  de52afb6-9bb2-4f1a-a3ce-4ef768c6a2fb     reviewed
#> 54  7e338671-08c7-4d7a-ac70-5b7bcaa46bec     reviewed
#> 55  49663870-adcd-4f18-b89f-187b71bdaadd     reviewed
#> 56  3f8f8bbe-d987-4797-a8ed-6bfeaf241963     reviewed
#> 57  8ea8893b-3403-4f8e-b68a-76e6517a9e45     reviewed
#> 58  ea0c3ca2-0f66-4062-ae8c-e4f9ea67b4f0     reviewed
#> 59  5243d831-1cc8-47c3-b222-cb8296faf0b5     reviewed
#> 60  9da20e2b-b8eb-42c7-b4a4-82f3627c3b9f     reviewed
#> 61  96fa3016-37cd-451d-9e0b-4fae1d5c078c     reviewed
#> 62  528c5e3b-c5f1-40df-9a36-8200b882634e     reviewed
#> 63  8b2681ef-8949-4fd9-80dd-209ee14bf7e1     reviewed
#> 64  00d0d3e7-f735-43ed-8711-b015888a94a6     reviewed
#> 65  bc771e37-0bab-4faa-866d-c78cac5eb02d     reviewed
#> 66  f54e1de8-8c71-45bf-bd05-49542135a42d     reviewed
#> 67  8ee9cf25-ece9-4c21-886c-23936be81f30     reviewed
#> 68  de8c243d-1a74-493c-bfe0-6e4d3a5e5573     reviewed
#> 69  9e27ca6f-bc01-4375-bd04-286b489ae45e     reviewed
#> 70  e1754a38-9168-4440-a756-ec694fc75b05     reviewed
#> 71  880797b7-56cb-48a8-8d41-9b052f031062     reviewed
#> 72  d2973148-2c7f-4bcb-9aee-72bc890bad52     reviewed
#> 73  8a650dac-691c-41cd-8690-5f160fd2a536     reviewed
#> 74  a771d726-5905-4ed6-b6f5-a61730e8b909     reviewed
#> 75  a829bde7-f971-4d88-8c3c-2d1e7a6d426a     reviewed
#> 76  b61f6af0-8934-4c07-b4cf-9907b10345f1     reviewed
#> 77  9f76d22e-da25-4285-ba97-88a424c1b44e     reviewed
#> 78  f6189f95-69d2-41f4-a2c7-e6b089b2660c     reviewed
#> 79  f775a113-abc2-403d-bc05-f01aa134f659     reviewed
#> 80  15814aac-7cd9-445d-86ae-014abeca9327     reviewed
#> 81  654b931c-8a6e-4b16-b752-2ffd1ecfa93a     reviewed
#> 82  bec2dab7-bd05-4d4b-8a9c-b2d5ee2e3990     reviewed
#> 83  b289a12a-8f05-48f7-a992-6218e867e4d9     reviewed
#> 84  f3885a3b-40a0-447d-b51c-29de72c11181     reviewed
#> 85  cabcf242-14a0-4ebd-a2a9-4a5a026fc47d     reviewed
#> 86  14c1a1eb-4154-476f-900b-b49767e398c3     reviewed
#> 87  c45ea1a5-6684-4199-859b-3dfcc8430f63     reviewed
#> 88  9fc7222b-056a-4764-bb2a-78e3437c691a     reviewed
#> 89  696c3086-db2b-4172-a2d2-3278be723312     reviewed
#> 90  306bc0b7-b0b1-421e-b1d4-32a2a9cd3e35     reviewed
#> 91  306d6d02-854e-4c9f-9aaf-a5ef7409cbf9     reviewed
#> 92  9b38552a-5c55-4032-9483-96615fa9c4b7     reviewed
#> 93  19a53ddd-7bbe-43b3-bd56-c1c35302d185     reviewed
#> 94  f39313dd-9885-45d2-b94d-c31004cb389b     reviewed
#> 95  869bb223-f969-44d9-ad28-25aff294be0a     reviewed
#> 96  7afd54ce-fad0-4f3d-a3f8-e05d9ec96eaa     reviewed
#> 97  314a846a-9db9-4530-bb39-bafe170f275e     reviewed
#> 98  40a69f9c-a867-46b1-88ac-c65d8918afa4     reviewed
#> 99  2a159df3-3c29-4e6b-838b-4b2427df4bbb     reviewed
#> 100 727547bb-3fdd-4b8b-b39b-43fb3a71f639     reviewed
#> 101 58023e5f-c934-4dc7-b674-be68d8503928     reviewed
#> 102 187c9c4e-6dfa-4668-983e-494488a5a192     reviewed
#> 103 d1ec9ff6-b3c2-4da3-b0bb-9af5d032da72     reviewed
#> 104 52acf2e6-4a7f-486d-b196-7a7a2918e318     reviewed
#> 105 d2061e68-bb20-4c3e-bc81-af4da5848eb2     reviewed
#> 106 36f3c2d6-096d-43e9-8220-3d37048ad8d5     reviewed
#> 107 b8cf9dd2-71c6-4ded-a20c-26b4013471c2     reviewed
#> 108 8d6a4cf0-73eb-4bf7-87cc-dbfe6dd94cf8     reviewed
#> 109 f9eb8cdf-b1cd-405c-b77f-f81242796b21     reviewed
#> 110 de7fac8c-9730-42f8-aa7f-7e394f3999ec     reviewed
#> 111 6749a1ae-9719-4c01-b64a-8dd0dbc70f89     reviewed
#> 112 1075bd7c-dc8d-443a-88ea-70b63f8f31c5     reviewed
#> 113 c7736013-0466-49ae-b75d-bc55c76fd25c     reviewed
#> 114 9508ba2f-26d2-422c-888d-da2de9fe6aca     reviewed
#> 115 339a458f-c7b6-458d-866a-1cff4bd68893     reviewed
#> 116 965c2cef-4e5a-4ab4-86c3-03004c270d1d     reviewed
#> 117 6bebb772-ff2c-494b-9ca2-6ed6cccc2a55     reviewed
#> 118 949af912-6e66-4f7b-b1eb-a53b970dbaad     reviewed
#> 119 573dadbb-6787-4a3b-bb3e-7c3fcd05601a     reviewed
#> 120 ca08ca74-e84c-4a51-b41b-fcd1f0736acd     reviewed
#> 121 f457f562-ad2a-4b1a-a5b0-8f0fc906882b     reviewed
#> 122 9f7fdadb-a88d-4dd1-b9fd-f2dcc75eae16     reviewed
#> 123 a506314e-2799-4bfe-ae7c-ee7cf1f51477     reviewed
#> 124 04d73e6d-b28c-4f0f-add7-f74b5aae4bd3     reviewed
#> 125 0d4a66fd-ca19-46df-a2a2-656a413e2aa3     reviewed
#> 126 f0da2abe-81be-4903-8868-cf7ca311e7de     reviewed
#> 127 7b3f3e72-4f35-42c8-893d-e29c6b474232     reviewed
#> 128 243a3ca3-77c8-4c34-9382-c1a656b16089     reviewed
#> 129 cf64687a-6494-418c-a7c2-8445f4abec32     reviewed
#> 130 220c80b8-532c-4f76-ac52-324ebfc91dd8     reviewed
#> 131 8e9ed9ab-2060-4fb8-85d1-5c3c71fd0da9     reviewed
#> 132 705172c3-aee0-4347-ba0a-f9c929be804c     reviewed
#> 133 234012f2-b78e-401e-964d-5fd7095d064e     reviewed
#> 134 be6dde0e-0660-42fb-afad-98ba6d20771c     reviewed
#> 135 be1c8637-b041-4ffd-ad0a-8f0758702009     reviewed
#> 136 6d8262ac-56ab-44ee-b2ca-fe958a7e7c1f     reviewed
#> 137 177f7ab6-041e-4f5c-a0cd-5f8c91a4a8a7     reviewed
#> 138 bc388e4a-ffc0-483b-9f36-25a04bc01711     reviewed
#> 139 02cf103e-a8a6-42a1-ba55-7481a23afeb6     reviewed
#> 140 202640d3-990d-4688-8965-1a653477bbc7     reviewed
#> 141 7dab441c-de7b-4dc1-a8ed-2bd756bec1e3     reviewed
#> 142 ae8cf93f-3e12-44c0-97c5-e6a3150e17a1     reviewed
#> 143 0e09c2f7-12fa-4f31-9ea1-3b3f03536c65     reviewed
#> 144 bf1cc03c-a78a-480d-9e69-92fd10bd112d     reviewed
#> 145 93299316-b92f-4462-9798-648a0b195860     reviewed
#> 146 5ddf2015-4930-47f9-836e-603214bffc7f     reviewed
#> 147 c3d1b8d5-9ffb-49b8-a1bf-bf3d7dfd1ac3     reviewed
#> 148 8cf6abe0-9899-4be8-9541-d5549e98ce46     reviewed
#> 149 8d9d13b6-ba90-4947-b2a0-9def78360574     reviewed
#> 150 0743ea67-5c03-4767-a910-b5e9a32e48f1     reviewed
#> 151 e1e506e8-8f98-4978-a6a1-e623920570e2     reviewed
#> 152 8ecbc153-9020-482a-a74c-74033c79367b     reviewed
#> 153 d5cbc3d1-caa0-4a7a-8303-a0687e5a6713     reviewed
#> 154 be160887-cca6-4650-beac-55b7c8417312     reviewed
#> 155 dcbf59e3-8c18-4bf3-af65-5aea763f4190     reviewed
#> 156 e05b82e2-a8d8-48ce-afc9-46ac84276733     reviewed
#> 157 e774a0f6-7d44-4f73-9bbb-728c07fd1027     reviewed
#> 158 dc7d3a3a-848d-4f13-bc41-68c9975ac26a     reviewed
#> 159 8482046d-dcb5-41f8-a2ef-9b2f725c5d84     reviewed
#> 160 542e5095-6f4c-4466-a077-5f9677430c78     reviewed
#> 161 f7dde74e-51da-46fc-bc64-9acc5fddcf68     reviewed
#> 162 348fe71b-84b0-4859-bf47-113a8c3321d9     reviewed
#> 163 30a6a2d9-f44a-4aba-8485-2f05ec0187a5     reviewed
#> 164 a0f15ce8-ccfb-4205-bbe3-745db91151b1     reviewed
#> 165 27a74d6d-9476-4eb3-a490-0a84dd5741d8     reviewed
#> 166 53ebe455-b051-4d68-b1f1-91c8987a2375     reviewed
#> 167 8dee3c07-4a29-4ec5-8053-e86af33a2676     reviewed
#> 168 2da0f8a8-d821-4369-b80f-3c2a3dfaad98     reviewed
#> 169 e60e771f-b24f-46cc-a5b6-afa5f094c2c6     reviewed
#> 170 2991da7d-ac03-4b79-bc32-516834dc79d0     reviewed
#> 171 67fb2172-566c-4e77-a3d4-dbda716703cb     reviewed
#> 172 31cf1894-9247-443d-b8ef-3008c864db27     reviewed
#> 173 c6ab15d2-2767-4887-bdf6-7eb170697b17     reviewed
#> 174 28c43949-28e0-4bc6-86e6-ab87a152ca95     reviewed
#> 175 db717f43-6492-44b4-8dbb-bdd7dad4eee9     reviewed
#> 176 175e1080-1ede-418b-863b-9462928d5d35     reviewed
#> 177 1d63d82b-a842-46e6-9aff-e8d778e944e5     reviewed
#> 178 369ca024-a9c7-4864-bd60-69e32f350aef     reviewed
#> 179 617498e5-121b-4666-bd4f-2272eb15a153     reviewed
#> 180 9878feef-3fba-42ef-a42d-157afb079c17     reviewed
#> 181 82f4bc9e-9099-4545-81cd-ac3a4277dd36     reviewed
#> 182 f6200457-9d45-4965-99ac-dc0878e6b849     reviewed
#> 183 6e4855f5-981a-40b5-9dcc-6fa3865a9d61     reviewed
#> 184 e260ef2b-3cf8-4220-a767-56c576b453f4     reviewed
#> 185 9fd06ffe-55ef-4709-88f9-6aa980c066b2     reviewed
#> 186 9ecf85fe-150c-42ce-be81-da863995fbc3     reviewed
#> 187 45e8887d-b4e6-4342-b851-6a635439e35e     reviewed
#> 188 eab1cec0-de2a-46c6-9c8d-2715331d9485     reviewed
#> 189 a8fc5340-aa47-4c8f-b354-7941b0cacd49     reviewed
#> 190 51e7bbdf-3d63-491a-97b2-8babec8192c0     reviewed
#> 191 f1aa12e8-f2be-4f22-918a-b3a2a13170d1     reviewed
#> 192 3ec3ecf5-c3b2-4820-8753-edfe57ef48a4     reviewed
#> 193 f2b97bad-43d5-43ec-a18d-12a122821501     reviewed
#> 194 fe65caef-757e-4124-84f4-6f844eb5bfac     reviewed
#> 195 0a0b9bb2-0c6a-4eea-b9a9-eae7c0f52440     reviewed
#> 196 2bfa6855-fed6-4e1a-b027-5be6986ce395     reviewed
#> 197 64661288-1ffe-41ec-9f82-218e30b063fe     reviewed
#> 198 4f571a2c-184b-45ae-a366-11c06ca5a778     reviewed
#> 199 dfd7ef49-bb9f-4747-88db-997d37a3b1ba     reviewed
#> 200 ae1b9d7e-63bf-40f4-80ba-351a5e6ea445     reviewed
#> 201 09407e6d-5b5e-4640-b553-af706ff37cd5     reviewed
#> 202 92f0b5f7-ea5f-4e00-adfa-182d86cab2f3     reviewed
#> 203 3afcd769-8903-45a8-a944-65769f41b8a3     reviewed
#> 204 6e08d5e6-7a82-48eb-85db-88abee91792f     reviewed
#> 205 32a13708-fbc2-453c-bcd2-a1be76d7b985     reviewed
#> 206 cff2596f-69df-40a3-9c32-57119277089c     reviewed
#> 207 033f7ed7-d7c1-4bbc-83be-8d89617f76fe     reviewed
#> 208 e5b1165d-7cec-4104-be15-075ac38a4163     reviewed
#> 209 b41738ce-c67e-414f-b048-00bd331d3bfb     reviewed
#> 210 94f145b0-9581-4e73-9c55-9ad4baf89198     reviewed
#> 211 53b41a8f-305f-417c-b321-85154d9b8a6d     reviewed
#> 212 b93bb1c3-5cc8-4232-bfb7-2b7896c8df6d     reviewed
#> 213 9986b407-f77b-4466-9c07-38c11d5da726     reviewed
#> 214 e144d01a-114e-4414-86b8-e7d60e7ce87c     reviewed
#> 215 deaf53bb-fc6e-4e85-bdc4-5434e65b06cd     reviewed
#> 216 16bf59ca-a54b-41f4-85ea-68dfa04b89a6     reviewed
#> 217 4a8a7825-5cea-4ac5-a717-e7bcd7bbf821     reviewed
#> 218 673c1604-4a14-45b9-bf2f-15f130d1d41e     reviewed
#> 219 9daec55c-eb02-4a7d-b4bb-4f0848a8a997     reviewed